Task-1/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
Werkzeug==3.0.1
